requirements.txt
absl-py==0.12.0
affine==2.3.0
amqp==5.0.5
asgiref==3.3.4
astunparse==1.6.3
attrs==20.3.0
billiard==3.6.3.0
boto3==1.17.111
botocore==1.20.111
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cligj==0.7.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
dateutils==0.6.12
defusedxml==0.7.1
dill==0.3.3
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-rest-framework-braces==0.3.4
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
drf-chunked-upload==0.4.2
drf-yasg==1.20.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
image==1.5.33
importlib-metadata==3.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kombu==5.0.2
Markdown==3.3.4
MarkupSafe==1.1.1
multiprocess==0.70.11.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==3.4.10.35
opt-einsum==3.3.0
packaging==20.9
piexif==1.1.3
Pillow==8.1.2
prompt-toolkit==3.0.16
protobuf==3.15.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pyproj==3.0.1
python-dateutil==2.8.1
python-xmp-toolkit==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
rasterio==1.2.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
scipy==1.5.4
six==1.15.0
snuggs==1.4.7
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xmp-tool==0.9.2
zipp==3.4.1
